requirements.txt

python-dotenv==0.21.1
django-allauth==0.52.0
django-ckeditor==6.5.1
celery==5.2.7
redis==4.4.0